Create Horn. Call among the Winds
An experimental soundscape inspired by the image of a mountain in the fog. The composition is created using oscillators and noise with Cauchy distribution and exponential distribution, which gives the sound variability and atmosphere. Smooth modulations and stereo effects create a feeling of wind and a distant call in the vast space.